requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2
whitenoise==6.2.0
django-location-field==2.1.0
drf-yasg==1.21.5
django_seed==0.3.1
django-filter==22.1
django-allauth==0.53.1
dj_rest_auth==3.0.0
phonenumbers==8.13.7
Pillow==9.4.0
django-phonenumber-field==7.0.2
uritemplate==4.1.1

aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
bootstrap4==0.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==40.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.6
django-allauth==0.54.0
django-crispy-forms==2.0
django-filter==23.1
django-location-field==2.7.0
django-phonenumber-field==7.0.2
django-seed==0.3.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.21.5
environs==8.0.0
Faker==18.3.4
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
mysql-connector-python==8.0.32
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
phonenumbers==8.13.8
Pillow==9.4.0
protobuf==3.20.3
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
scipy==1.10.1
six==1.16.0
sketchpy==0.1.0
source==1.2.0
sqlparse==0.2.4
svg.path==6.2
svgpathtools==1.6.0
svgwrite==1.4.3
toposort==1.10
tqdm==4.64.1
turtle==0.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0
yarl==1.8.2
